+++ app/components/groupe_gestionnaire/groupe_gestionnaire_gestionnaires/gestionnaire_component.rb
@@ -1,9 +1,10 @@
 class GroupeGestionnaire::GroupeGestionnaireGestionnaires::GestionnaireComponent < ApplicationComponent
   include ApplicationHelper
 
-  def initialize(groupe_gestionnaire:, gestionnaire:)
+  def initialize(groupe_gestionnaire:, gestionnaire:, is_gestionnaire: true)
     @groupe_gestionnaire = groupe_gestionnaire
     @gestionnaire = gestionnaire
+    @is_gestionnaire = is_gestionnaire
   end
 
   def email
@@ -23,7 +24,7 @@ class GroupeGestionnaire::GroupeGestionnaireGestionnaires::GestionnaireComponent
   end
 
   def remove_button
-    if is_there_at_least_another_active_admin?
+    if is_there_at_least_another_active_admin? && @is_gestionnaire
       button_to 'Retirer',
        gestionnaire_groupe_gestionnaire_gestionnaire_path(@groupe_gestionnaire, @gestionnaire),
        method: :delete,
